Add ScoreTextSystem
This is so that changes to the score made in the editor show up in game more quickly.
