+++ examples/pong/systems/winner.rs
@@ -2,8 +2,7 @@ use amethyst::assets::AssetStorage;
 use amethyst::audio::output::Output;
 use amethyst::audio::Source;
 use amethyst::core::transform::Transform;
-use amethyst::ecs::prelude::{Entity, Join, Read, ReadExpect, System, Write, WriteStorage};
-use amethyst::ui::UiText;
+use amethyst::ecs::prelude::{Join, Read, ReadExpect, System, Write, WriteStorage};
 use audio::Sounds;
 use {Ball, ScoreBoard};
 
@@ -16,11 +15,9 @@ impl<'s> System<'s> for WinnerSystem {
     type SystemData = (
         WriteStorage<'s, Ball>,
         WriteStorage<'s, Transform>,
-        WriteStorage<'s, UiText>,
         Write<'s, ScoreBoard>,
         Read<'s, AssetStorage<Source>>,
         ReadExpect<'s, Sounds>,
-        ReadExpect<'s, ScoreText>,
         Option<Read<'s, Output>>,
     );
 
@@ -29,11 +26,9 @@ impl<'s> System<'s> for WinnerSystem {
         (
             mut balls,
             mut transforms,
-            mut text,
             mut score_board,
             storage,
             sounds,
-            score_text,
             audio_output,
         ): Self::SystemData,
     ) {
@@ -45,16 +40,10 @@ impl<'s> System<'s> for WinnerSystem {
             let did_hit = if ball_x <= ball.radius {
                 // Right player scored on the left side.
                 score_board.score_right += 1;
-                if let Some(text) = text.get_mut(score_text.p2_score) {
-                    text.text = score_board.score_right.to_string();
-                }
                 true
             } else if ball_x >= ARENA_WIDTH - ball.radius {
                 // Left player scored on the right side.
                 score_board.score_left += 1;
-                if let Some(text) = text.get_mut(score_text.p1_score) {
-                    text.text = score_board.score_left.to_string();
-                }
                 true
             } else {
                 false
@@ -80,10 +69,4 @@ impl<'s> System<'s> for WinnerSystem {
             }
         }
     }
-}
-
-/// Stores the entities that are displaying the player score with UiText.
-pub struct ScoreText {
-    pub p1_score: Entity,
-    pub p2_score: Entity,
 }
